+++ src/main/java/com/example/application/helpers/validators/ElectricSensorValidator.java
@@ -2,7 +2,7 @@ package com.example.application.helpers.validators;
 
 public class ElectricSensorValidator {
 
-    public static boolean validateLowHigh(float low, float high) {
+    public static boolean validateLowHigh(Double low, Double high) {
         return low < high;
     }
